+++ __tests__/_apiIntegrationHelpers.js
@@ -86,152 +86,6 @@ export function getOrgItems(orgs, org, csvLines) {
   }
 }
 
-const cleanReaddirSync = (path) => {
-  let cleanDirectories = [];
-
-  if (fs.existsSync(path)) {
-    cleanDirectories = fs.readdirSync(path)
-      .filter((file) => file !== '.DS_Store');
-  } else {
-    console.warn(`no such file or directory, ${path}`);
-  }
-
-  return cleanDirectories;
-};
-
-const EMPTY_TIME = '0001-01-01T00:00:00+00:00';
-
-/**
- *
- * @param resourcesPath
- * @return {null|*[]}
- */
-export const getLocalResourceList = (resourcesPath = USER_RESOURCES_PATH) => {
-  try {
-    if (!fs.existsSync(resourcesPath)) {
-      fs.ensureDirSync(resourcesPath);
-    }
-
-    const localResourceList = [];
-    const resourceLanguages = fs.readdirSync(resourcesPath)
-      .filter((file) => path.extname(file) !== '.json' && file !== '.DS_Store');
-
-    for (let i = 0; i < resourceLanguages.length; i++) {
-      const languageId = resourceLanguages[i];
-      const biblesPath = path.join(resourcesPath, languageId, 'bibles');
-      const tHelpsPath = path.join(resourcesPath, languageId, TRANSLATION_HELPS);
-      const bibleIds = cleanReaddirSync(biblesPath);
-      const tHelpsResources = cleanReaddirSync(tHelpsPath);
-
-      bibleIds.forEach((bibleId) => {
-        const bibleIdPath = path.join(biblesPath, bibleId);
-        const bibleLatestVersion = getLatestVersion(bibleIdPath);
-
-        if (bibleLatestVersion) {
-          const pathToBibleManifestFile = path.join(bibleLatestVersion, 'manifest.json');
-
-          if (fs.existsSync(pathToBibleManifestFile)) {
-            const resourceManifest = fs.readJsonSync(pathToBibleManifestFile);
-            const remoteModifiedTime = (resourceManifest.remoteModifiedTime !== EMPTY_TIME) && resourceManifest.remoteModifiedTime;
-            const localResource = {
-              languageId: languageId,
-              resourceId: bibleId,
-              version: path.base(bibleLatestVersion, true),
-              modifiedTime: remoteModifiedTime || resourceManifest.catalog_modified_time,
-            };
-
-            localResourceList.push(localResource);
-          } else {
-            console.warn(`no such file or directory, ${pathToBibleManifestFile}`);
-          }
-        } else {
-          console.log(`$bibleLatestVersion is ${bibleLatestVersion}.`);
-        }
-      });
-
-      tHelpsResources.forEach((tHelpsId) => {
-        const tHelpResource = path.join(tHelpsPath, tHelpsId);
-        const tHelpsLatestVersion = getLatestVersion(tHelpResource);
-
-        if (tHelpsLatestVersion) {
-          const pathTotHelpsManifestFile = path.join(tHelpsLatestVersion, 'manifest.json');
-
-          if (fs.existsSync(pathTotHelpsManifestFile)) {
-            const resourceManifest = fs.readJsonSync(pathTotHelpsManifestFile);
-            const localResource = {
-              languageId: languageId,
-              resourceId: tHelpsId,
-              modifiedTime: resourceManifest.catalog_modified_time,
-            };
-
-            localResourceList.push(localResource);
-          } else {
-            console.warn(`no such file or directory, ${pathTotHelpsManifestFile}`);
-          }
-        } else {
-          console.log(`$tHelpsLatestVersion is ${tHelpsLatestVersion}.`);
-        }
-      });
-    }
-    return localResourceList;
-  } catch (error) {
-    console.error(error);
-    return null;
-  }
-};
-
-/**
- * Returns the versioned folder within the directory with the highest value.
- * e.g. `v10` is greater than `v9`
- * @param {string} dir - the directory to read
- * @return {string} the full path to the latest version directory.
- */
-export function getLatestVersion(dir) {
-  const versions = listVersions(dir);
-
-  if (versions.length > 0) {
-    return path.join(dir, versions[0]);
-  } else {
-    return null;
-  }
-}
-
-/**
- * Returns an array of paths found in the directory filtered and sorted by version
- * @param {string} dir
- * @return {string[]}
- */
-function listVersions(dir) {
-  if (fs.pathExistsSync(dir)) {
-    const versionedDirs = fs.readdirSync(dir).filter((file) => fs.lstatSync(path.join(dir, file)).isDirectory() &&
-      file.match(/^v\d/i));
-    return versionedDirs.sort((a, b) =>
-      -compareVersions(a, b), // do inverted sort
-    );
-  }
-  return [];
-}
-
-/**
- * compares version numbers, if a > b returns 1; if a < b return -1; else are equal and return 0
- * @param a
- * @param b
- * @return {number}
- */
-export function compareVersions(a, b) {
-  const cleanA = semver.coerce(a);
-  const cleanB = semver.coerce(b);
-
-  if (semver.gt(cleanA, cleanB)) {
-    return 1;
-  } else if (semver.lt(cleanA, cleanB)) {
-    return -1;
-  } else {
-    return 0;
-  }
-}
-
-
 /**
  * @description This helper method generates a timestamp in milliseconds for use
  *              in the storing of data in the app. Timestamps will be used to
